# perf: Non blocking GenServer.handle_call