--- a/lib/nif.ex
+++ b/lib/nif.ex
@@ -158,19 +158,6 @@ defmodule Mediasoup.Nif do
   def plain_transport_event(_transport, _pid, _event_types),
     do: :erlang.nif_error(:nif_not_loaded)
 
-  ## consumer with async
-  defp consumer_get_stats_async(_consumer), do: :erlang.nif_error(:nif_not_loaded)
-  defp consumer_pause_async(_consumer), do: :erlang.nif_error(:nif_not_loaded)
-  defp consumer_resume_async(_consumer), do: :erlang.nif_error(:nif_not_loaded)
-
-  defp consumer_set_preferred_layers_async(_consumer, _referred_layers),
-    do: :erlang.nif_error(:nif_not_loaded)
-
-  defp consumer_set_priority_async(_consumer, _priority), do: :erlang.nif_error(:nif_not_loaded)
-  defp consumer_unset_priority_async(_consumer), do: :erlang.nif_error(:nif_not_loaded)
-  defp consumer_request_key_frame_async(_consumer), do: :erlang.nif_error(:nif_not_loaded)
-  defp consumer_dump_async(_consumer), do: :erlang.nif_error(:nif_not_loaded)
-
   ## producer with async
   defp producer_pause_async(_producer), do: :erlang.nif_error(:nif_not_loaded)
   defp producer_resume_async(_producer), do: :erlang.nif_error(:nif_not_loaded)
@@ -389,27 +376,22 @@ defmodule Mediasoup.Nif do
   def consumer_preferred_layers(_consumer), do: :erlang.nif_error(:nif_not_loaded)
   def consumer_current_layers(_consumer), do: :erlang.nif_error(:nif_not_loaded)
 
-  def consumer_get_stats(consumer),
-    do: consumer_get_stats_async(consumer) |> handle_async_nif_result() |> unwrap_ok()
+  def consumer_get_stats_async(_consumer, _from), do: :erlang.nif_error(:nif_not_loaded)
 
-  def consumer_pause(consumer), do: consumer_pause_async(consumer) |> handle_async_nif_result()
-  def consumer_resume(consumer), do: consumer_resume_async(consumer) |> handle_async_nif_result()
+  def consumer_pause_async(_consumer, _from), do: :erlang.nif_error(:nif_not_loaded)
+  def consumer_resume_async(_consumer, _from), do: :erlang.nif_error(:nif_not_loaded)
 
-  def consumer_set_preferred_layers(consumer, referred_layers),
-    do:
-      consumer_set_preferred_layers_async(consumer, referred_layers) |> handle_async_nif_result()
+  def consumer_set_preferred_layers_async(_consumer, _referred_layers, _from),
+    do: :erlang.nif_error(:nif_not_loaded)
 
-  def consumer_set_priority(consumer, priority),
-    do: consumer_set_priority_async(consumer, priority) |> handle_async_nif_result()
+  def consumer_set_priority_async(_consumer, _priority, _from),
+    do: :erlang.nif_error(:nif_not_loaded)
 
-  def consumer_unset_priority(consumer),
-    do: consumer_unset_priority_async(consumer) |> handle_async_nif_result()
+  def consumer_unset_priority_async(_consumer, _from), do: :erlang.nif_error(:nif_not_loaded)
 
-  def consumer_request_key_frame(consumer),
-    do: consumer_request_key_frame_async(consumer) |> handle_async_nif_result()
+  def consumer_request_key_frame_async(_consumer, _from), do: :erlang.nif_error(:nif_not_loaded)
 
-  def consumer_dump(consumer),
-    do: consumer_dump_async(consumer) |> handle_async_nif_result() |> unwrap_ok()
+  def consumer_dump_async(_consumer, _from), do: :erlang.nif_error(:nif_not_loaded)
 
   # data_consumer
   @spec data_consumer_id(reference) :: String.t()
@@ -499,6 +481,7 @@ defmodule Mediasoup.Nif do
     end
   end
 
-  defp unwrap_ok({:ok, result}), do: result
-  defp unwrap_ok(result), do: result
+  def unwrap_ok({:ok, {}}), do: {:ok}
+  def unwrap_ok({:ok, result}), do: result
+  def unwrap_ok(result), do: result
 end