--- a/lib/nif.ex
+++ b/lib/nif.ex
@@ -56,22 +56,6 @@ defmodule Mediasoup.Nif do
   ## webrtc_server with async
   defp webrtc_server_dump_async(_transport), do: :erlang.nif_error(:nif_not_loaded)
 
-  ## pipe_transport with async
-  defp pipe_transport_get_stats_async(_transport), do: :erlang.nif_error(:nif_not_loaded)
-  defp pipe_transport_dump_async(_transport), do: :erlang.nif_error(:nif_not_loaded)
-  defp pipe_transport_consume_async(_transport, _option), do: :erlang.nif_error(:nif_not_loaded)
-  defp pipe_transport_connect_async(_transport, _option), do: :erlang.nif_error(:nif_not_loaded)
-  defp pipe_transport_produce_async(_transport, _option), do: :erlang.nif_error(:nif_not_loaded)
-
-  defp pipe_transport_set_max_incoming_bitrate_async(_transport, _bitrate),
-    do: :erlang.nif_error(:nif_not_loaded)
-
-  defp pipe_transport_consume_data_async(_transport, _option),
-    do: :erlang.nif_error(:nif_not_loaded)
-
-  defp pipe_transport_produce_data_async(_transport, _option),
-    do: :erlang.nif_error(:nif_not_loaded)
-
   # plain transport
   ## properties
   def plain_transport_tuple(_transport), do: :erlang.nif_error(:nif_not_loaded)
@@ -265,36 +249,32 @@ defmodule Mediasoup.Nif do
   @spec pipe_transport_closed(reference) :: boolean
   def pipe_transport_closed(_transport), do: :erlang.nif_error(:nif_not_loaded)
 
-  def pipe_transport_consume(transport, option),
-    do: pipe_transport_consume_async(transport, option) |> handle_async_nif_result()
+  def pipe_transport_consume_async(_transport, _option, _from),
+    do: :erlang.nif_error(:nif_not_loaded)
 
-  def pipe_transport_consume_data(transport, option),
-    do: pipe_transport_consume_data_async(transport, option) |> handle_async_nif_result()
+  def pipe_transport_consume_data_async(_transport, _option, _from),
+    do: :erlang.nif_error(:nif_not_loaded)
 
-  def pipe_transport_connect(transport, option),
-    do: pipe_transport_connect_async(transport, option) |> handle_async_nif_result()
+  def pipe_transport_connect_async(_transport, _option, _from),
+    do: :erlang.nif_error(:nif_not_loaded)
 
-  def pipe_transport_produce(transport, option),
-    do: pipe_transport_produce_async(transport, option) |> handle_async_nif_result()
+  def pipe_transport_produce_async(_transport, _option, _from),
+    do: :erlang.nif_error(:nif_not_loaded)
 
-  def pipe_transport_set_max_incoming_bitrate(transport, bitrate),
-    do:
-      pipe_transport_set_max_incoming_bitrate_async(transport, bitrate)
-      |> handle_async_nif_result()
+  def pipe_transport_set_max_incoming_bitrate_async(_transport, _bitrate, _from),
+    do: :erlang.nif_error(:nif_not_loaded)
 
-  def pipe_transport_produce_data(transport, option),
-    do: pipe_transport_produce_data_async(transport, option) |> handle_async_nif_result()
+  def pipe_transport_produce_data_async(_transport, _option, _from),
+    do: :erlang.nif_error(:nif_not_loaded)
 
-  def pipe_transport_get_stats(transport),
-    do: pipe_transport_get_stats_async(transport) |> handle_async_nif_result() |> unwrap_ok()
+  def pipe_transport_get_stats_async(_transport, _from), do: :erlang.nif_error(:nif_not_loaded)
 
   def pipe_transport_tuple(_transport), do: :erlang.nif_error(:nif_not_loaded)
   def pipe_transport_sctp_parameters(_transport), do: :erlang.nif_error(:nif_not_loaded)
   def pipe_transport_sctp_state(_transport), do: :erlang.nif_error(:nif_not_loaded)
   def pipe_transport_srtp_parameters(_transport), do: :erlang.nif_error(:nif_not_loaded)
 
-  def pipe_transport_dump(transport),
-    do: pipe_transport_dump_async(transport) |> handle_async_nif_result() |> unwrap_ok()
+  def pipe_transport_dump_async(_transport, _from), do: :erlang.nif_error(:nif_not_loaded)
 
   def pipe_transport_event(_transport, _pid, _event_types),
     do: :erlang.nif_error(:nif_not_loaded)