--- a/lib/webrtc_server.ex
+++ b/lib/webrtc_server.ex
@@ -112,6 +112,7 @@ defmodule Mediasoup.WebRtcServer do
     GenServer.start_link(__MODULE__, %{reference: reference}, opt)
   end
 
+  @impl true
   def init(state) do
     Process.flag(:trap_exit, true)
     {:ok, supervisor} = DynamicSupervisor.start_link(strategy: :one_for_one)