## Benchmark 
Benchmark our product(private repo) produce/consume 
### before
```
Name                      ips        average  deviation         median         99th %
produce/consume          1.37      727.91 ms    ±26.72%      734.94 ms     1119.05 ms
```
### after
```
Name                      ips        average  deviation         median         99th %
produce/consume          1.75      571.91 ms    ±11.26%      556.91 ms      697.27 ms
```

<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

- **New Features**
	- Enhanced asynchronous handling for various media operations, improving responsiveness.
	- Introduced a `TermBox` struct for thread-safe term management.
	- Added new test cases to cover WebRTC server lifecycle management.
	- Introduced new parameters for improved context handling in asynchronous operations across multiple modules.

- **Bug Fixes**
	- Improved error handling and response management for asynchronous operations.

- **Documentation**
	- Updated comments and specifications to reflect changes in function signatures and behavior.

- **Chores**
	- Removed unnecessary imports and deprecated modules to streamline codebase.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->